requirements.txt
opencv-python
numpy
streamlit
Pillow
matplotlib
scikit-image
pandas
scikit-learn
streamlit-cropper
tensorflow
seaborn
cv2
